Only overwrite extract.sh if build ID matches
If we put extract.sh in the outer archive, it will always be replaced
even if the build ID check fails. So if a user tries to install an old
archive, we might actually roll back their extract.sh script. :O

Instead, make sure the _inner_ archive always has the latest copy, and
extract it manually before installing the rest of the inner archive.