More update script updates for more better updating (#9)
* Default FROM_BUILDID to FROM_COMMIT

For future releases (both full and incremental), the build ID will just
be the abbreviated Git commit hash from the internal repo, so we can set
the build ID check automatically.

For the initial 2410.2 patch, we have multiple base build IDs, but we
can override: FROM_BUILDID='472b162|a10951c4|d1bae326' ./gen_update.sh

Also, for repeated testing, can do FROM_BUILDID='*' to effectively
disable the check :)

* Copy latest extract.sh into outer update archive

This means we'll always extract the inner archive using the latest code.
It also means we'll need to be a little bit careful about staging
changes to extract.sh going forward, but we can always remove or change
this logic, so I think that's okay.

* No need to delete update.sh on build ID mismatch

extract.sh already takes care of this for us.

* Don't try to recursively remove archive file

It's a file, so -r does nothing, and removing it is slightly safer
(e.g., don't accidentally remove all of /opt if ARCHIVE_NAME is somehow
unset).

* Create .done file after installing inner archive

We reboot at the end of update.sh, so we never return to extract.sh and
it never marks the archive as done.

* Simplify sed commands

* Only overwrite extract.sh if build ID matches

If we put extract.sh in the outer archive, it will always be replaced
even if the build ID check fails. So if a user tries to install an old
archive, we might actually roll back their extract.sh script. :O

Instead, make sure the _inner_ archive always has the latest copy, and
extract it manually before installing the rest of the inner archive.

* Bootstrap inclusion of pv for update progress